pyproject.toml
# SPDX-FileCopyrightText: 2022 Magenta ApS <https://magenta.dk>
# SPDX-License-Identifier: MPL-2.0

[tool.poetry]
name = "elevate-manager"
version = "0.0.0"  # versioning handled by conventional commits
description = "OS2mo AMQP trigger receiver that elevates manager engagement according to manager level changes"
authors = ["Magenta ApS <info@magenta.dk>"]
readme = "README.md"
packages = [{include = "elevate_manager"}]
homepage = "https://magenta.dk/"
repository = "https://git.magenta.dk/rammearkitektur/os2mo-triggers/os2mo-amqp-trigger-elevate-manager"
keywords = ["os2mo", "amqp"]

[tool.poetry.dependencies]
python = "^3.11"
structlog = "^24"
pydantic = "^1.10.4"
more-itertools = "^9.0.0"
uvicorn = "^0.23"
fastramqpi = "^10"
pytest-cov = "^4.0.0"
gql = "^3.4.0"
click = "^8.1.3"
pytest-asyncio = "^0.21.0"
pytest-split = "^0.8.1"
ariadne-codegen = {version = "^0.7.1", extras = ["subscriptions"]}

[tool.poetry.group.pre-commit.dependencies]
mypy = "^1"
pre-commit = "^3"
ruff = "^0.4"
deptry = "^0.16"

[tool.ariadne-codegen]
target_package_name = "autogenerated_graphql_client"
target_package_path = "elevate_manager/"
client_name = "GraphQLClient"
schema_path = "schema.graphql" # Found by running "curl -O http://localhost:5000/graphql/v22/schema.graphql"
queries_path = "queries.graphql"
plugins = [
    "ariadne_codegen.contrib.shorter_results.ShorterResultsPlugin",
]
[tool.ariadne-codegen.scalars.DateTime]
type = "datetime.datetime"
[tool.ariadne-codegen.scalars.UUID]
type = "uuid.UUID"

[tool.poetry.group.dev.dependencies]

[tool.pytest.ini_options]
asyncio_mode="auto"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
